Test the accuracy of uTR and RepeatMasker
This pipeline tests the accuracy of uTR and RepeatMasker using benchmark data randomly generated from mosaic tandem repeat seed patterns. C programs in subdirectories must be compiled in advance. The pipeline generates benchmark data, applies uTR to the benchmark, checks the accuracy of uTR, applies RepeatMasker to the data, and tests the accuracy of RepeatMasker.