Comparison between uTR and RepeatMasker
Since RepeatMasker is not designed to predict mosaic tandem repeats, it might be unfair to use it for this purpose. With this in mind, this table should be examined. This table compares the accuracy and wall clock times of uTR and RepeatMasker processing various benchmark data sets generated from different patterns of mosaic tandem repeats. "tmp.zip" shows the many files output by uTR and RepeatMasker in this analysis.